Moved nodal files t root drectory
