--- dbserver/.env
@@ -1,1 +1,0 @@
-EXAMPLE_ENV_VAR=hello
