--- dbserver/.npmignore
@@ -1,5 +1,0 @@
-.DS_Store
-.env
-.polybit
-node_modules/
-npm-debug.log
