--- dbserver/README.md
@@ -1,11 +1,0 @@
-# A Nodal Application
-
-## Running Locally
-
-```sh
-nodal s
-```
-
-Your app should now be running on [localhost:3000](http://localhost:3000/).
-
-Generated by [Nodal](http://nodaljs.com)
