--- dbserver/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "nodalDemo2",
-  "version": "0.0.0",
-  "description": "nodalDemo2",
-  "author": "titanium_pinewood",
-  "main": "server.js",
-  "engines": {
-    "node": "6.0.0"
-  },
-  "dependencies": {
-    "bcrypt": "^0.8.6",
-    "nodal": "0.10.0"
-  },
-  "devDependencies": {
-    "chai": "~3.5.0",
-    "mocha": "~2.4.5"
-  },
-  "scripts": {
-    "test": "mocha ./test/runner.js",
-    "start": "node cluster.js",
-    "worker": "node worker.js"
-  },
-  "private": true
-}
